Move `isidentifier` to hy.lex
